Work around 'cannot chown: permission denied'.
Using systemd-tmpfiles itself to create the basic folder
structure is nifty, but it has the clear disadvantage that
while creating the buildroot, it's not running as root and
thus 'chown' will fail.

This caused people before to omit the owner/group parts in
filesystem.conf, which would be fine... except we ship this
file on every system, and, on top of that, we ship more files
that *do* have owner/group bits for these folders.

This causes a 'duplicate entry' warning to be emitted every
time tmpfiles.d activates, since two entries are found to be
*conflicting*.

However, if two entries are found to be *identical*, then
tmpfiles.d doesn't emit any warning messages.

The proper solution therefore should be that (1) filesystem.conf
should not provide half-baked tmpfiles.d definitions (and, we
can fix ownership bits in the .spec %files section anyway), and
(2), we should somehow just omit using `fchown()` during the
%install section.

As systemd-tmpfiles.d does not have a mode of operation where
you can omit ownership change bits, I've opted to replace the
usage of systemd-tmpfiles with 5 lines of shell code that
achieve the same effect.

This then allows us to solve the conflicting tmpfiles.d definitions
in the way one would expect it to be solved.