Install a default locale.conf
Support a stateless locale.conf, located in the /usr/share/defaults/etc
directory.

The en_US.UTF-8 locale is already initialized by
locale-archive-trigger.service, but to make it enabled everywhere,
export the supported environment variables to login shells in the
stateless /etc/profile. This fixes the locale variables, for example, in
ssh shells.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>