Update prompt for bash and zsh
This is a part of a bigger plan that bourne compatible shells shall
have a same set of default profiles.

This patch shall be applied on `50-prompt.sh` in
`clearlinux-pkgs/filesystem`, it includes the following:

1. It would be set the prompt correctly for both bash and zsh.
   Currently the prompt for bash is defined here, but the prompt for
   zsh is set in `/usr/share/defaults/. Moreover, their prompts are
   not identical.
2. The color for bash prompt is not set correctly, it's fixed here.
3. The end character for zsh is not set correctly, as well as the
   color, which are both fixed now.
4. When the terminal is xterm*, this file sets the title for both
   bash and zsh. Currently the prompt for bash is defined here, but
   it's not set for zsh.
5. Refactored variable names to increase readability.

I also wrap the code in functions so and made the variable local.
Also, I used the hooks in bash and zsh.

This patch will requires the changes in the following:

1. `profile.x86_64` and `profile.i386` in `clearlinux-pkgs/filesystem`
   shall not set anything related to prompt.
2. `0001-stateless-configuration.patch` in `clearlinux-pkgs/zsh`
   shall be patched so that `zshrc` won't set the prompt.

This has been tested locally for sh (aliased to bash), bash, and zsh.