Ensure interactive terminal has rows and columns
Fixes: https://github.com/clearlinux/clr-installer/issues/475

As of 29850, we are now using /usr/bin/login from util-linux instead
of /usr/bin/login from shadow package. This change resulted in not
properly initializing the terminal rows and columns on some consoles
like QEMU.

cut and setterm were used as they are a part os-core and to ensure
we support both bash and zsh.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>