Don't run `stty` and friends if input is not a terminal.
We have a bunch of reports showing error output like this:

```
stty: 'standard input': Inappropriate ioctl for device
stty: 'standard input': Inappropriate ioctl for device
setterm: stdin does not refer to a terminal
```

These are caused by running `stty` on things that are not a tty.

This is easily avoided entirely, in the above way.