Temporary workaround
Due to an issue with rebuilds of clr-init causing systems to be unable
to boot if they were using LUKS coinciding with the removal of the
libgfortran-avx package (which clr-init had a BuildRequires on) add a
temporary provides for libgfortran-avx.

filesystem was chosen as something that won't be new and it is fast to
rebuild. Remove this provides once a new clr-init is released that
resolves this issue.